+++ internal/provider/repository/format/docker.go
@@ -81,6 +81,15 @@ func (f *DockerRepositoryFormatHosted) DoReadRequest(state any, apiClient *sonat
 
 	// Call to API to Read
 	apiResponse, httpResponse, err := apiClient.RepositoryManagementAPI.GetDockerHostedRepository(ctx, stateModel.Name.ValueString()).Execute()
+
+	// Temporary Workaround:
+	// latest_policy not returned from READ API for Docker Hosted
+	if stateModel.Storage.LatestPolicy.IsNull() {
+		apiResponse.Storage.LatestPolicy = common.NewFalse()
+	} else {
+		apiResponse.Storage.LatestPolicy = stateModel.Storage.LatestPolicy.ValueBoolPointer()
+	}
+
 	return *apiResponse, httpResponse, err
 }