+++ internal/provider/repository/repository_common.go
@@ -434,7 +434,9 @@ func hostedStandardSchema(repoFormat string, repoType format.RepositoryType) tfs
 	// LatestPolicy is only for Docker Hosted Repositories
 	if repoFormat == common.REPO_FORMAT_DOCKER && repoType == format.REPO_TYPE_HOSTED {
 		storageAttributes["latest_policy"] = schema.ResourceOptionalBoolWithDefault(
-			"Whether to allow redeploying the 'latest' tag but defer to the Deployment Policy for all other tags. Only applicable for Hosted Docker Repositories when Deployment Policy is set to Disable.",
+			`Whether to allow redeploying the 'latest' tag but defer to the Deployment Policy for all other tags. Only applicable for Hosted Docker Repositories when Deployment Policy is set to Disable.
+
+  **NOTE:** The APIs for Sonatype Nexus Repository do not currently allow for reading this value - hence during import this may result in terraform plan changes as we defualt to false.`,
 			false,
 		)
 	}