+++ internal/provider/repository/repository_npm_x_resources_test.go
@@ -157,6 +157,8 @@ resource "%s" "repo" {
 					resource.TestCheckResourceAttr(resourceProxyName, "http_client.authentication.password", "pass"),
 					resource.TestCheckResourceAttr(resourceProxyName, "http_client.authentication.preemptive", "true"),
 					resource.TestCheckResourceAttr(resourceProxyName, "http_client.authentication.type", "username"),
+					resource.TestCheckResourceAttr(resourceProxyName, "replication.preemptive_pull_enabled", "false"),
+					resource.TestCheckNoResourceAttr(resourceProxyName, "replication.asset_path_regex"),
 					resource.TestCheckNoResourceAttr(resourceProxyName, "routing_rule"),
 					resource.TestCheckResourceAttr(resourceProxyName, "npm.remove_quarrantined", "true"),